@@ -9,7 +9,7 @@ description = "Modern, async-first Python client generator for OpenAPI specifica
 authors = [{ name = "Mindhive Oy", email = "contact@mindhive.fi" }]
 maintainers = [{ name = "Ville Venäläinen | Mindhive Oy", email = "ville@mindhive.fi" }]
 readme = "README.md"
-requires-python = ">=3.10,<4.0.0"
+requires-python = ">=3.12,<4.0.0"
 dependencies = [
   "openapi-spec-validator>=0.7",
   "openapi-core>=0.19",
@@ -44,8 +44,6 @@ classifiers = [
   "Operating System :: POSIX :: Linux",
   "Operating System :: Microsoft :: Windows",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11", 
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3 :: Only",
   "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
