@@ -9,9 +9,10 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 from pyopenapi_gen.core.utils import NameSanitizer
 from pyopenapi_gen.core.writers.code_writer import CodeWriter
+from pyopenapi_gen.types.strategies.response_strategy import ResponseStrategy
 
 # Import necessary helpers from endpoint_utils if they were used directly or indirectly
-from pyopenapi_gen.helpers.endpoint_utils import get_param_type, get_return_type_unified  # Added
+from pyopenapi_gen.helpers.endpoint_utils import get_param_type  # Removed get_return_type_unified - now use strategy
 
 if TYPE_CHECKING:
     from pyopenapi_gen import IROperation
@@ -32,13 +33,15 @@ class EndpointMethodSignatureGenerator:
         op: IROperation,
         context: RenderContext,
         ordered_params: List[Dict[str, Any]],
+        strategy: ResponseStrategy,
     ) -> None:
         """Writes the method signature to the provided CodeWriter."""
         # Logic from EndpointMethodGenerator._write_method_signature
         for p_info in ordered_params:  # Renamed p to p_info to avoid conflict if IRParameter is named p
             context.add_typing_imports_for_type(p_info["type"])
 
-        return_type = get_return_type_unified(op, context, self.schemas)
+        # Use strategy return type instead of computing it again
+        return_type = strategy.return_type
         context.add_typing_imports_for_type(return_type)
 
         # Check if AsyncIterator is in return_type or any parameter type
