@@ -1,10 +1,10 @@
 """
-Unit tests for PythonConstructRenderer JSON Wizard functionality.
+Unit tests for PythonConstructRenderer BaseSchema functionality.
 
 Scenario: Test the PythonConstructRenderer's ability to generate dataclasses
-with JSONWizard inheritance and field mapping configuration.
+with BaseSchema inheritance and field mapping configuration.
 
-Expected Outcome: Generated code includes JSONWizard inheritance and proper
+Expected Outcome: Generated code includes BaseSchema inheritance and proper
 field mapping configuration when needed.
 """
 
@@ -12,13 +12,13 @@ from pyopenapi_gen.context.render_context import RenderContext
 from pyopenapi_gen.core.writers.python_construct_renderer import PythonConstructRenderer
 
 
-class TestPythonConstructRendererJsonWizard:
-    """Test PythonConstructRenderer JSON Wizard functionality."""
+class TestPythonConstructRendererBaseSchema:
+    """Test PythonConstructRenderer BaseSchema functionality."""
 
-    def test_render_dataclass__no_field_mappings__generates_standard_dataclass(self):
+    def test_render_dataclass__no_field_mappings__generates_base_schema_dataclass(self) -> None:
         """
         Scenario: Render a dataclass without any field mappings.
-        Expected Outcome: Standard dataclass without JSONWizard inheritance.
+        Expected Outcome: BaseSchema dataclass without Meta class.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -35,18 +35,18 @@ class TestPythonConstructRendererJsonWizard:
         )
 
         # Assert
-        assert "class User:" in result
-        assert "JSONWizard" not in result
-        assert "key_transform_with_load" not in result
+        assert "class User(BaseSchema):" in result
+        assert "with automatic JSON field mapping" in result
+        assert "key_transform_with_load" not in result  # No Meta class when no mappings
         assert "@dataclass" in result
         assert "name: str" in result
         assert "age: int" in result
         assert "email: Optional[str] = None" in result
 
-    def test_render_dataclass__with_field_mappings__generates_json_wizard_dataclass(self):
+    def test_render_dataclass__with_field_mappings__generates_base_schema_dataclass(self) -> None:
         """
         Scenario: Render a dataclass with field mappings.
-        Expected Outcome: Dataclass with JSONWizard inheritance and Meta class.
+        Expected Outcome: Dataclass with BaseSchema inheritance and Meta class.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -56,7 +56,7 @@ class TestPythonConstructRendererJsonWizard:
             ("last_name", "str", None, "Last name (maps from 'lastName')"),
             ("id_", "str", None, "User ID (maps from 'id')"),
         ]
-        field_mappings = {"firstName": "first_name", "lastName": "last_name", "id": "id_"}
+        field_mappings: dict[str, str] = {"firstName": "first_name", "lastName": "last_name", "id": "id_"}
 
         # Act
         result = renderer.render_dataclass(
@@ -68,18 +68,18 @@ class TestPythonConstructRendererJsonWizard:
         )
 
         # Assert
-        assert "class User(JSONWizard):" in result
+        assert "class User(BaseSchema):" in result
         assert "with automatic JSON field mapping" in result
-        assert "class _(JSONWizard.Meta):" in result
+        assert "class Meta:" in result
         assert "key_transform_with_load = {" in result
         assert "'firstName': 'first_name'," in result
         assert "'id': 'id_'," in result
         assert "'lastName': 'last_name'," in result
 
-    def test_render_dataclass__empty_field_mappings__generates_standard_dataclass(self):
+    def test_render_dataclass__empty_field_mappings__generates_base_schema_dataclass(self) -> None:
         """
         Scenario: Render a dataclass with empty field mappings dict.
-        Expected Outcome: Standard dataclass without JSONWizard inheritance.
+        Expected Outcome: BaseSchema dataclass without Meta class.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -97,14 +97,14 @@ class TestPythonConstructRendererJsonWizard:
         )
 
         # Assert
-        assert "class User:" in result
-        assert "JSONWizard" not in result
-        assert "key_transform_with_load" not in result
+        assert "class User(BaseSchema):" in result
+        assert "with automatic JSON field mapping" in result
+        assert "key_transform_with_load" not in result  # No Meta class for empty mappings
 
-    def test_render_dataclass__with_field_mappings__adds_proper_imports(self):
+    def test_render_dataclass__with_field_mappings__adds_proper_imports(self) -> None:
         """
         Scenario: Render a dataclass with field mappings and check imports.
-        Expected Outcome: Both dataclass and JSONWizard imports are added.
+        Expected Outcome: Both dataclass and BaseSchema imports are added.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -125,13 +125,13 @@ class TestPythonConstructRendererJsonWizard:
         imports = context.import_collector.imports
         assert "dataclasses" in imports
         assert "dataclass" in imports["dataclasses"]
-        assert "dataclass_wizard" in imports
-        assert "JSONWizard" in imports["dataclass_wizard"]
+        assert "..core.schemas" in imports
+        assert "BaseSchema" in imports["..core.schemas"]
 
-    def test_render_dataclass__no_field_mappings__only_dataclass_import(self):
+    def test_render_dataclass__no_field_mappings__includes_base_schema_import(self) -> None:
         """
-        Scenario: Render a standard dataclass and check imports.
-        Expected Outcome: Only dataclass import is added, no JSONWizard import.
+        Scenario: Render a dataclass without field mappings and check imports.
+        Expected Outcome: Both dataclass and BaseSchema imports are added.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -145,9 +145,10 @@ class TestPythonConstructRendererJsonWizard:
         imports = context.import_collector.imports
         assert "dataclasses" in imports
         assert "dataclass" in imports["dataclasses"]
-        assert "dataclass_wizard" not in imports
+        assert "..core.schemas" in imports
+        assert "BaseSchema" in imports["..core.schemas"]
 
-    def test_render_dataclass__sorted_field_mappings__generates_deterministic_output(self):
+    def test_render_dataclass__sorted_field_mappings__generates_deterministic_output(self) -> None:
         """
         Scenario: Render a dataclass with unsorted field mappings.
         Expected Outcome: Mappings are sorted in the output for consistency.
@@ -173,10 +174,10 @@ class TestPythonConstructRendererJsonWizard:
         assert a_field_pos != -1
         assert z_field_pos != -1
 
-    def test_render_dataclass__no_fields__generates_empty_class_without_json_wizard(self):
+    def test_render_dataclass__no_fields__generates_empty_class_with_base_schema(self) -> None:
         """
         Scenario: Render a dataclass with no fields.
-        Expected Outcome: Empty dataclass with pass statement, no JSONWizard.
+        Expected Outcome: Empty BaseSchema dataclass with pass statement.
         """
         # Arrange
         renderer = PythonConstructRenderer()
@@ -186,11 +187,11 @@ class TestPythonConstructRendererJsonWizard:
         result = renderer.render_dataclass(class_name="Empty", fields=[], description="Empty class", context=context)
 
         # Assert
-        assert "class Empty:" in result
+        assert "class Empty(BaseSchema):" in result
         assert "pass" in result or "No properties defined in schema" in result
-        assert "JSONWizard" not in result
+        assert "with automatic JSON field mapping" in result
 
-    def test_render_dataclass__single_field_mapping__generates_meta_class(self):
+    def test_render_dataclass__single_field_mapping__generates_meta_class(self) -> None:
         """
         Scenario: Render a dataclass with a single field mapping.
         Expected Outcome: Meta class is generated even for single mapping.
@@ -207,6 +208,6 @@ class TestPythonConstructRendererJsonWizard:
         )
 
         # Assert
-        assert "class User(JSONWizard):" in result
-        assert "class _(JSONWizard.Meta):" in result
+        assert "class User(BaseSchema):" in result
+        assert "class Meta:" in result
         assert "'userId': 'user_id'," in result
